EloquentJS 20: fileserver public space
Add fileserver file to new project with frontend client that allows
editing of the very website being served to the user. There is a basic
editor class that builds the DOM and maintains some simple state. Then a
few methods help control the file selection, loading a file or
directory's contents, and updating the select box or text area. Finally,
when the form is submitted, the client sends a PUT request to the
fileserver with the new contents.